# openapi.yaml
openapi: 3.0.3
info:
  title: API Estoque Vitória Cestas
  version: 1.2.0
  description: Endpoints para gestão de itens, fornecedores, autenticação e demais cadastros previstos no banco de dados.
servers:
  - url: http://{host}
    variables:
      host:
        default: localhost:3000
  - url: https://projeto-vitoriacestas-backend.vercel.app
paths:
...
  /api/addresses:
    get:
      summary: Lista endereços cadastrados
      tags: [Endereços]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Endereços retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  addresses:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/EnderecoInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra endereço
      tags: [Endereços]
      security:
        - bearerAuth: []
      description: Cria um novo endereço para ser associado a clientes ou fornecedores.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnderecoInput'
      responses:
        '201':
          description: Endereço criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnderecoResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/customers:
    get:
      summary: Lista clientes cadastrados
      tags: [Clientes]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Clientes retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  customers:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/ClienteInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra cliente
      tags: [Clientes]
      security:
        - bearerAuth: []
      description: Cria um cliente com CPF ou CNPJ, data de nascimento e endereço obrigatório.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
      responses:
        '201':
          description: Cliente criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
        '409':
          description: CPF ou CNPJ duplicado
  /api/raw-materials:
    get:
      summary: Lista matérias-primas
      tags: [Matérias-primas]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Materiais retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  materials:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/MateriaPrimaInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra matéria-prima
      tags: [Matérias-primas]
      security:
        - bearerAuth: []
      description: Registra uma matéria-prima com informações de custo, validade e características.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MateriaPrimaInput'
      responses:
        '201':
          description: Matéria-prima criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MateriaPrimaResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/materials:
    get:
      summary: Lista matérias-primas (alias de /api/raw-materials)
      tags: [Matérias-primas]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Materiais retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  materials:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/MateriaPrimaInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra matéria-prima (alias de /api/raw-materials)
      tags: [Matérias-primas]
      security:
        - bearerAuth: []
      description: Registra uma matéria-prima com informações de custo, validade e características.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MateriaPrimaInput'
      responses:
        '201':
          description: Matéria-prima criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MateriaPrimaResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/manufacturing:
    get:
      summary: Lista relações de manufatura
      tags: [Manufatura]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Relações retornadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  manufacturing:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/ManufaturaInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Registra composição de manufatura
      tags: [Manufatura]
      security:
        - bearerAuth: []
      description: Define a relação entre produto e matéria-prima utilizada, com a quantidade consumida.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufaturaInput'
      responses:
        '201':
          description: Relação de manufatura criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManufaturaResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/manufaturas:
    get:
      summary: Lista relações de manufatura (alias)
      tags: [Manufatura]
      security:
        - bearerAuth: []
      description: Alias português para /api/manufacturing, disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Relações retornadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  manufacturing:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/ManufaturaInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Registra composição de manufatura (alias)
      tags: [Manufatura]
      security:
        - bearerAuth: []
      description: Disponível em português como alias de /api/manufacturing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufaturaInput'
      responses:
        '201':
          description: Relação de manufatura criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManufaturaResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
...
  /api/entregas-material:
    get:
      summary: Lista entradas de matéria-prima (alias)
      tags: [Entregas de Material]
      security:
        - bearerAuth: []
      description: Alias em português para /api/material-deliveries.
      responses:
        '200':
          description: Entradas retornadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  deliveries:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/EntregaMaterialInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra entrada de matéria-prima (alias)
      tags: [Entregas de Material]
      security:
        - bearerAuth: []
      description: Endpoint em português equivalente a /api/material-deliveries.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntregaMaterialInput'
      responses:
        '201':
          description: Entrada de material registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntregaMaterialResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/orders:
    get:
      summary: Lista pedidos
      tags: [Pedidos]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Pedidos retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/PedidoInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra pedido
      tags: [Pedidos]
      security:
        - bearerAuth: []
      description: Cria um pedido para um cliente, com dados do presenteado quando aplicável.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoInput'
      responses:
        '201':
          description: Pedido criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/pedidos:
    get:
      summary: Lista pedidos (alias)
      tags: [Pedidos]
      security:
        - bearerAuth: []
      description: Alias em português para /api/orders.
      responses:
        '200':
          description: Pedidos retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/PedidoInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra pedido (alias)
      tags: [Pedidos]
      security:
        - bearerAuth: []
      description: Endpoint em português equivalente a /api/orders.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoInput'
      responses:
        '201':
          description: Pedido criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/product-shipments:
    get:
      summary: Lista envios de produtos
      tags: [Envios]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Envios retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  shipments:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/EnvioProdutoInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Registra envio de produto
      tags: [Envios]
      security:
        - bearerAuth: []
      description: Liga um pedido a um produto enviado, com data e preço aplicados.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvioProdutoInput'
      responses:
        '201':
          description: Envio registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvioProdutoResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/envios:
    get:
      summary: Lista envios de produtos (alias)
      tags: [Envios]
      security:
        - bearerAuth: []
      description: Alias em português para /api/product-shipments.
      responses:
        '200':
          description: Envios retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  shipments:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/EnvioProdutoInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Registra envio de produto (alias)
      tags: [Envios]
      security:
        - bearerAuth: []
      description: Endpoint em português equivalente a /api/product-shipments.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvioProdutoInput'
      responses:
        '201':
          description: Envio registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvioProdutoResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/feedbacks:
    get:
      summary: Lista feedbacks
      tags: [Feedbacks]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Feedbacks retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  feedbacks:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/FeedbackInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra feedback
      tags: [Feedbacks]
      security:
        - bearerAuth: []
      description: Registra a avaliação de um cliente sobre um pedido ou serviço prestado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackInput'
      responses:
        '201':
          description: Feedback registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/feedback:
    get:
      summary: Lista feedbacks (alias)
      tags: [Feedbacks]
      security:
        - bearerAuth: []
      description: Alias singular para /api/feedbacks.
      responses:
        '200':
          description: Feedbacks retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  feedbacks:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/FeedbackInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra feedback (alias)
      tags: [Feedbacks]
      security:
        - bearerAuth: []
      description: Endpoint singular equivalente a /api/feedbacks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackInput'
      responses:
        '201':
          description: Feedback registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
  /api/phones:
    get:
      summary: Lista telefones de clientes
      tags: [Telefones]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Telefones retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  phones:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/TelefoneInput'
                        - type: object
                          properties:
                            id:
                              type: integer
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra telefone de cliente
      tags: [Telefones]
      security:
        - bearerAuth: []
      description: Associa um telefone adicional a um cliente existente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelefoneInput'
      responses:
        '201':
          description: Telefone cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelefoneResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EnderecoInput:
      type: object
      required: [rua, cep, numero]
      properties:
        rua:
          type: string
        cep:
          type: string
        numero:
          type: string
    EnderecoResponse:
      type: object
      properties:
        message:
          type: string
        address:
          allOf:
            - $ref: '#/components/schemas/EnderecoInput'
            - type: object
              properties:
                id:
                  type: integer
    ClienteInput:
      type: object
      required: [nome, dataNascimento, enderecoId]
      properties:
        cnpj:
          type: string
        cpf:
          type: string
        nome:
          type: string
        email:
          type: string
          format: email
        dataNascimento:
          type: string
          format: date
        enderecoId:
          type: integer
          nullable: false
    ClienteResponse:
      type: object
      properties:
        message:
          type: string
        customer:
          allOf:
            - $ref: '#/components/schemas/ClienteInput'
            - type: object
              properties:
                id:
                  type: integer
    MateriaPrimaInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
        tipo:
          type: string
        custo:
          type: number
          format: float
          nullable: true
          default: 0
        dataValidade:
          type: string
          format: date
        descricao:
          type: string
        tamanho:
          type: string
        material:
          type: string
        acessorio:
          type: string
    MateriaPrimaResponse:
      type: object
      properties:
        message:
          type: string
        material:
          allOf:
            - $ref: '#/components/schemas/MateriaPrimaInput'
            - type: object
              properties:
                id:
                  type: integer
    ManufaturaInput:
      type: object
      required: [produtoId, materialId, quantidadeMaterial]
      properties:
        produtoId:
          type: integer
        materialId:
          type: integer
        quantidadeMaterial:
          type: integer
          default: 0
          minimum: 0
    ManufaturaResponse:
      type: object
      properties:
        message:
          type: string
        manufacturing:
          allOf:
            - $ref: '#/components/schemas/ManufaturaInput'
            - type: object
              properties:
                id:
                  type: integer
    EntregaMaterialInput:
      type: object
      required: [materialId, quantidade, dataEntrada]
      properties:
        materialId:
          type: integer
        fornecedorId:
          type: integer
          nullable: true
          default: null
        quantidade:
          type: integer
          default: 0
          minimum: 0
        dataEntrada:
          type: string
          format: date
        custo:
          type: number
          format: float
          nullable: true
          default: 0
    EntregaMaterialResponse:
      type: object
      properties:
        message:
          type: string
        delivery:
          allOf:
            - $ref: '#/components/schemas/EntregaMaterialInput'
            - type: object
              properties:
                id:
                  type: integer
    PedidoInput:
      type: object
      required: [clienteId, preco, dataPedido]
      properties:
        clienteId:
          type: integer
          nullable: true
        endereco:
          type: string
        preco:
          type: number
          format: float
          default: 0
          minimum: 0
        dataPedido:
          type: string
          format: date
        cpfPresentado:
          type: string
        nomePresentado:
          type: string
        emailPresentado:
          type: string
          format: email
        enderecoPresentado:
          type: string
    PedidoResponse:
      type: object
      properties:
        message:
          type: string
        order:
          allOf:
            - $ref: '#/components/schemas/PedidoInput'
            - type: object
              properties:
                id:
                  type: integer
...